+++ waspc/ChangeLog.md
@@ -28,7 +28,6 @@ Follow the [the official migration guide](https://wasp.sh/docs/migration-guides/
 ### 🎉 New Features
 
 - Wasp now supports `onAfterEmailVerified` auth hooks! You can use this hook to run custom logic after a user has verified their email.
-- Auth: you can now use Slack as a social auth provider (by @scorpil)
 
 ### 🐞 Bug fixes
 
@@ -40,6 +39,12 @@ Follow the [the official migration guide](https://wasp.sh/docs/migration-guides/
 - Show a friendlier error when there are no routes defined in the wasp file ([#2643](https://github.com/wasp-lang/wasp/pull/2643))
 - Modernized our TypeScript support for bundlers ([#2656](https://github.com/wasp-lang/wasp/pull/2656))
 - Added support for jump-to-definition for Wasp symbols ([#2656](https://github.com/wasp-lang/wasp/pull/2656))
+
+## 0.16.4
+
+### 🎉 New Features and improvements
+
+- Auth: you can now use Slack as a social auth provider (by @scorpil)
 
 ## 0.16.3